Move HTTP handlers to separate file.
This has no other purpose than making the code easier to navigate.
