+++ cmd/test-server/main.go
@@ -5,10 +5,18 @@ package main
 
 import (
 	"crypto/tls"
-	"encoding/json"
 	"errors"
 	"flag"
 	"fmt"
+	"log"
+	"math"
+	"net"
+	"net/http"
+	"os"
+	"path"
+	"strconv"
+	"time"
+
 	"github.com/go-ping/ping"
 	"github.com/google/gopacket"
 	"github.com/google/gopacket/layers"
@@ -16,14 +24,6 @@ import (
 	"github.com/google/uuid"
 	"github.com/gorilla/websocket"
 	"golang.org/x/net/ipv4"
-	"html/template"
-	"log"
-	"net"
-	"net/http"
-	"os"
-	"path"
-	"strconv"
-	"time"
 )
 
 const (
@@ -98,44 +98,6 @@ func isValidUUID(u string) bool {
 // fmtTimeMs returns the value (time.Duration) in milliseconds, the inbuilt time.Milliseconds() function only returns an int64 value
 func fmtTimeMs(value time.Duration) float64 {
 	return (float64(value) / float64(time.Millisecond))
-}
-
-// echoHandler for the echo webserver that speaks WebSocket
-func echoHandler(w http.ResponseWriter, r *http.Request) {
-	if checkHTTPParams(w, r, "/echo") {
-		return
-	}
-	c, err := upgrader.Upgrade(w, r, nil)
-	if err != nil {
-		ErrLogger.Println("upgrade:", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-
-	}
-	defer c.Close()
-	for {
-		mt, message, err := c.ReadMessage()
-		if err != nil {
-			ErrLogger.Println("read:", err)
-			break
-		}
-		// ReadMessage() returns messageType int, p []byte, err error]
-		var wsData map[string]interface{}
-		json.Unmarshal(message, &wsData)
-		if wsData["type"] != "ws-latency" {
-			if wsUUID, ok := wsData["UUID"].(string); ok {
-				// Only log the final message with all latencies calculated, and don't log other unsolicited echo messages
-				if isValidUUID(string(wsUUID)) {
-					InfoLogger.Println(string(message))
-				}
-			}
-		}
-		err = c.WriteMessage(mt, message)
-		if err != nil {
-			ErrLogger.Println("write:", err)
-			break
-		}
-	}
 }
 
 // getHeaderFromICMPResponsePayload parses IP headers from ICMP Response Payload of the icmpPkt and returns IP header, and error if any
@@ -392,41 +354,6 @@ func start0trace(uuid string, netConn net.Conn) map[int]HopRTT {
 	return traceroute
 }
 
-// traceHandler speaks WebSocket for extracting underlying connection to use for 0trace
-func traceHandler(w http.ResponseWriter, r *http.Request) {
-	if checkHTTPParams(w, r, "/trace") {
-		return
-	}
-	var uuid string
-	for k, v := range r.URL.Query() {
-		if k == "uuid" {
-			uuid = v[0]
-		}
-	}
-
-	c, err := upgrader.Upgrade(w, r, nil)
-	if err != nil {
-		ErrLogger.Println("upgrade:", err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	defer c.Close()
-	myConn := c.UnderlyingConn()
-	traceroute := start0trace(uuid, myConn)
-	results := TracerouteResults{
-		UUID:      uuid,
-		Timestamp: time.Now().UTC().Format("2006-01-02T15:04:05.000000"),
-		HopData:   traceroute,
-	}
-	zeroTraceResult, err := json.Marshal(results)
-	if err != nil {
-		ErrLogger.Println("Error logging 0trace results: ", err)
-		InfoLogger.Println(results) // Dump results in non-JSON format
-	}
-	zeroTraceString := string(zeroTraceResult)
-	InfoLogger.Println(zeroTraceString)
-}
-
 // IcmpPinger sends ICMP pings and returns statistics
 func IcmpPinger(ip string) RtItem {
 	pinger, err := ping.NewPinger(ip)
@@ -475,48 +402,6 @@ func checkHTTPParams(w http.ResponseWriter, r *http.Request, pathstring string) 
 		return true
 	}
 	return false
-}
-
-// pingHandler for ICMP measurements which also serves the webpage via a template
-func pingHandler(w http.ResponseWriter, r *http.Request) {
-	if checkHTTPParams(w, r, "/ping") {
-		return
-	}
-	clientIPstr := r.RemoteAddr
-	clientIP, _, _ := net.SplitHostPort(clientIPstr)
-
-	// Concurrently send ICMP pings for a <batchSizeLimit> number of IPs
-	var icmpResults []RtItem
-	icmpResults = append(icmpResults, IcmpPinger(clientIP))
-
-	// Combine all results
-	results := Results{
-		UUID:   uuid.NewString(),
-		IPaddr: clientIP,
-		//RFC3339 style UTC date time with added seconds information
-		Timestamp:   time.Now().UTC().Format("2006-01-02T15:04:05.000000"),
-		IcmpPing:    icmpResults,
-		AvgIcmpStat: getMeanIcmpRTT(icmpResults),
-	}
-
-	jsObj, err := json.Marshal(results)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	resultString := string(jsObj)
-	InfoLogger.Println(resultString)
-	var WebTemplate, _ = template.ParseFiles(path.Join(directoryPath, "pingpage.html"))
-	WebTemplate.Execute(w, results)
-}
-
-// indexHandler serves the default index page with reasons for scanning IPs on this server and point of contact
-func indexHandler(w http.ResponseWriter, r *http.Request) {
-	if checkHTTPParams(w, r, "/") {
-		return
-	}
-	path := path.Join(directoryPath, "/index.html")
-	http.ServeFile(w, r, path)
 }
 
 // redirectToTLS helps redirect HTTP connections to HTTPS
